refactor download and upload scripts
