+++ cerebrum/example/agents/academic_agent/config.json
@@ -9,7 +9,7 @@
     ],
     "meta": {
         "author": "example",
-        "version": "1.1.2",
+        "version": "1.1.5",
         "license": "CC0"
     },
     "build": {
