+++ cerebrum/script.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+# Run local agent
+python cerebrum/run_agent.py \
+  --mode local \
+  --path cerebrum/example/agents/test_agent \
+  --agenthub_url https://app.aios.foundation \
+  --task "Tell me what is the capital of United States"
